+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod db;
 pub mod server;
+pub(crate) mod services;
